+++ ee/packages/api-client/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @rocket.chat/api-client
 
+## 0.1.20
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.5.2
+- @rocket.chat/rest-typings@6.5.2
+
+## 0.1.19
+
+### Patch Changes
+
+- Updated dependencies [c2b224fd82]
+- Updated dependencies [c2b224fd82]
+  - @rocket.chat/rest-typings@6.5.1
+  - @rocket.chat/core-typings@6.5.1
+
 ## 0.1.18
 
 ### Patch Changes