+++ packages/ui-client/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @rocket.chat/ui-client
 
+## 3.0.2
+
+### Patch Changes
+
+- @rocket.chat/ui-contexts@3.0.2
+
+## 3.0.1
+
+### Patch Changes
+
+- @rocket.chat/ui-contexts@3.0.1
+
 ## 3.0.0
 
 ### Patch Changes